// src/types/partials/MedicationAdministrationDosage.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TQuantity } from '../partials/Quantity';
import { TRatio } from '../partials/Ratio';

export type TMedicationAdministrationDosage = {
    id?: String;
    extension?: TExtension[];
    modifierExtension?: TExtension[];
    text?: String;
    site?: TCodeableConcept;
    route?: TCodeableConcept;
    method?: TCodeableConcept;
    dose?: TQuantity;
    rateRatio?: TRatio;
    rateQuantity?: TQuantity;
};